Extract CommandPlanner interface
Why:
The CommandPlanner takes the List of classified Context's and plans the
next action for the roller. Actions it can decide to take are:

1. stop the roll. Everything is healthy or it's too degraded to take action.
2. delay. Wait some time and try again.
3. process a context.

So it is responsible for detecting when the roll should be terminated and
selecting the next context to act on. So a good class to implement the logic
about rolling the controller last and choosing appropriate action if rolling
a pod will break min-isr availability.