Prepare for Kafka Roller reimplementation
This paves the way for a KafkaRoller reimplementation. The idea is that we can
incrementally build the state machine Kafka Roller implementation by:

1. invoking a rolling restart with the new implementation
2. invoking the old implementation

The new implementation can handle more and more of the roll until the existing
KafkaRoller is a no-op because the system is brought to stable before it is
invoked.

Signed-off-by: Robert Young <robeyoun@redhat.com>